refactor(parser): extract reference scanners into scan submodule
